refactor: root init/destroy -> root/init.c
