+++ winbuild/Makefile
@@ -93,6 +93,7 @@ SRCS=\
 	root/crawler.c       \
 	root/dir.c       \
 	root/file.c       \
+	root/init.c       \
 	root/iothread.c       \
 	root/lock.c       \
 	root/notifythread.c       \
